HTML API: Guard against non-string attribute values.
Because there are three ways for an attribute in HTML to exist, the HTML API
reports three kinds of values for `get_attribute()` calls:
 - `null` means that no attribute exists of the given name
 - `true` means that the attribute exists but there is no value, e.g. '<input checked>'.
 - a string value means that the attribute exists and has a value, e.g. '<img src="test">'.

When operating on the value returned by `get_attribute()` then it's important to
ensure that it's a string value before treating it as one. A call to `empty()` is not
enough because a boolean attribute, being `true`, does not return `false` for `empty()`.

In this patch blocks that read and then use attribute values as strings have been updated
in order to guard against cases where the attribute might not be the string the code
expects.