Blocks: Skip unnecessary esc_attr() in HTML API call in Social Link block.
Becasue the HTML API already escapes values passed into it, it's not only
unnecessary to manually wrap attribute values in `esc_attr()`, but can be
confusing and lead to mixed-escaped or double-escaped values. It's better
to avoid the extra work and leave things plain as PHP values.

This patch removes the call to `esc_attr()` and replaces it with `trim()`,
which not only removes the extra work but also ensures that the `rel`
values don't have leading or trailing whitespace.