docs: Update E2E test documentation (#55519)
* docs: Rewrite end-to-end testing documentation

Update for breaking changes from upgrading Appium 2. Reduce verbosity
and duplication to improve the approachability and maintainability of
the documentation.

* docs: Remove E2E test contributing documentation

This was replaced by a section in the directory's README documentation.

* docs: Rewrite end-to-end testing Getting Started section

Update for breaking changes from upgrading Appium 2. Reduce verbosity
and duplication between the `WordPress/gutenberg` repository to improve
the approachability and maintainability of the documentation.

* docs: Update links in Gutenberg for Mobile Apps code documentation

Documents have been updated and consolidated to reduce verbosity
and duplication to improve the approachability and maintainability of
the documentation.