List Block: Hide appender
When the list block itself is selected the appender displays in the lower corner.
This can be surprising and distracting when writing in the list, and sometimes
cause visual glitches. This appender is not present on the list item block, so
the appearance and disappearance can be disorienting when moving the selection
throughout a list.

In this patch the appender is unconditionally disabled on the list block, which was
reportedly how it was supposed to have always been.

Co-authored-by: Gage Peterson <justgage@hey.com>
Co-authored-by: Ella <4710635+ellatrix@users.noreply.github.com>