test: E2E test setup script creates iOS simulators (#55716)
* refactor: Convert spaces to tabs to match project

Tabs are the project configuration's preferred format.

* test: E2E test script detects or create iOS simulators

Simplify E2E test environment setup by creating the required iOS
simulators if they are not present.

* test: E2E setup script throws error if required iOS runtime unavailable

Provide helpful directions to fix E2E test environment setup.

* docs: Update E2E test environment setup documentation

Adjust directions to account for the expanded capabilities of the
`test:e2e:setup` script.

* refactor: Abbreviate E2E test setup script logs

An attempt to improve comprehension through brevity.

* refactor: Avoid repetition

* feat: Colorize E2E test script log

Improve notability of errors.

* fix: Throw error when iOS runtime check returns empty response

The logic used to search for the runtime results in an empty response if
the runtime is not installed. Therefore, we need to verify the response
it not empty.