Remove header slot in favor of rendering BlockContextualToolbar directly
The slot in the header was important for rendering the block popover within the header, but we are no longer doing that. As such, we should render the BlockContextualToolbar directly in the header to simplify the implementation. Since we can't rely on forwarding a ref to the navigableToolbar from within the `<BlockTools />`, we needed to check if the escape keypress was coming from within a toolbar. If it is, we ignore the keypress to allow the block toolbar to handle it. This means we could remove all of the forwardRefs from the block toolbar.