Move clientId key to BlockContextualToolbar
The client id was set as a key on the SelectedBlockTools, which helped keep the block to block toolbar connection. This benefit wasn't applied to the BlockContextualToolbar, so fixed toolbars didn't get this bug fix. By moving this key down a level to the BlockContextualToolbar component, all the versions of the toolbar get this bug fix.