Package `@ariakit/test` should be a dev dependency
In #54939 a new test was added that depends on `@ariakit/test`. But,
even though that package is not needed in the production code, it was
(accidentally?) added to `dependencies` instead of `devDependencies`.

This means everyone using `@wordpress/components` winds up having to
install this package. Let's move it to devDependencies where it belongs.