fix remove version paramerer from deprecated calls
Till we have a agreed upon strategy for dealing with deprecations in WordPress core it doesn’t make sense to have incorrect version numbers logged.

Looking at the removed versions you can see that a good chunk of these is already in the past so they are just outright incorrect.