Large refactor of BlockToolbar rendering flow.
- Split out Block Toolbar Breadcrumb and Block Toolbar Popover
The Block Toolbar Popover and Block Toolbar Breadcrumb did not rely on each other and the logic of each other was intertwined in an unhelpful way. Split these out to have more control over the rendering of the toolbar.

- Remove initialIndex and focusing on mount for BlockToolbar
Now the BlockToolbar is always rendered if there are block tools for it, but in the case of the popover, it is hidden via CSS if we don't want it to show. This makes the logic simpler, removes quite a bit of code, and reduces renders. We now get the initial focus on mount index "for free" because we never unmount the component. The index resets automatically when the block changes.

-