Add raw handling tests
This file is a clone of the same `blocks-raw-handling.js` file located in `gutenberg/test/integration`. The reason for the separation is that several of the test cases fail in the native version. For now, we are going to skip them, but we'd need to work on them in the future.
Once all issues in tests are addressed, we'll remove this file in favor of the original one.