Update innerBlocks docs on defaultBlock api and provide example
I was incorrect when I wrote the original doc here and cast this as an array, its in fact an object. Also, provided a quick code example and a gif of this concept in action. Not sure what the rules are on including images/gifs in Gutenberg docs but thought it might help illustrate what defaultBlock enables.