[RNMobile] Fix HTML to blocks conversion when no transformations are available (#56723)
* Add native workaround for HTML block in `htmlToBlocks`

* Add raw handling tests

This file is a clone of the same `blocks-raw-handling.js` file located in `gutenberg/test/integration`. The reason for the separation is that several of the test cases fail in the native version. For now, we are going to skip them, but we'd need to work on them in the future.
Once all issues in tests are addressed, we'll remove this file in favor of the original one.

* Update blocks raw handling test snapshot with original values

* Disable more pasteHandler test cases due to not matching test snapshot

* Comment obsolete snapshots of blocks raw handling tests

The reason for commenting them instead of removing is that, in the future, we'll restore them once we address the failing test cases.