fix: Guard against a block styles crash due to null block values
In certain scenarios, e.g., when the editor hangs due to poor
performance, the `getBlock` value unexpectedly returns `null`. To guard
against a crash in this scenario, we now conditionally access attributes
on the block.

A ideal resolution would be improving editor performance to avoid the
scenario where a race condition causes a crash, but the fact remains
that `getBlock` _can_ return `null` values and that we should not
presume that it will not.