Do not mount block toolbar popover until needed.
Once the block toolbar popover is mounted, it will remain mounted until the block id changes. This is a lot of extra work to delay this mounting. It could be a performance improvement in one sense, but the top toolbar and mobile toolbars get remounted each time a block changes, so I lean towards removing all this work and keeping the code simpler.