Fix: Update Interactivity API Use of Private HTML API Classes to avoid Crash
Previously the HTML API was using a mixture of approaches in its helper classes
for describing spans of text. One was (start, end) and the other was (start, length).
In WordPress/wordpress-develop#5721 these were all normalized to (start, length) but
the Interactivity API was not updated in concert with that change.

In this patch the `inner_html` methods are updated to use the appropriate formats.

A following PR is necessary to update the compat layer so that crashes don't appear
on sites running older versions of WordPress with this updated code.