Ensure currentFontSize is always a number
Some automated tests are failing, reporting that "fontSize={NaN}" is being added as a parameter to some blocks. Ref: https://github.com/WordPress/gutenberg/actions/runs/7216537798/job/19662900342?pr=57055

I found this is happening in cases where "currentFontSize" is "undefined", which this new function had not accounted for. With this commit, a check is added, so we continue to return "undefined" when there is no number associated with "currentFontSize".