If the lock button has shown in the toolbar, keep it in the toolbar until unmounted.
If the block lock button has been shown, we don't want to remove it from the toolbar until the toolbar is rendered again without it. Removing it beforehand can cause focus loss issues, such as when unlocking the block from the modal. It needs to return focus from whence it came, and to do that, we need to leave the button in the toolbar.