Disable lock button if user cannot control lock state
The canLockBlocks setting allows a user to manage the block locked state. Admins always have it set to true, but some users are not allowed to unlock blocks. If they can't unlock a block, there is nothing in the toolbar that suggests a block is locked as there is for admins. I think it's useful to have some kind of indicator that the block is locked, so if a block is locked but a user cannot edit it, show a disabled lock icon in the toolbar.