Exclude disabled buttons when setting initial focus on toolbar
When the first item was disabled, it was causing the last item of the toolbar to be focused because the the toolbar was trying to focus a disabled button. This updates the check to only include focusable buttons so we do not try to send focus to an unfocusable button.