feat: Revert logic intended to display progress bar when offline
I believe this logic erroneously prevented the spinner progress bar from
displaying when an image is successfully uploading. I.e., the logic read
"if show spinner and online, apply 'progress bar hidden' styles." I
imagine this confusion is a result of the previous logic using the `||`
pattern to conditionally apply the hidden styles, which is confusing
in itself.

This logic should likely be implemented by the way the host app
communicates the upload state. Implementing this in the JavaScript logic
is confusing or misleading, in my opinion. It adds a new, somewhat
surprising interpretation of the existing states provided by the host
app.