feat: Hide writing prompt unless block is selected
Avoid displaying subsequent empty paragraphs to better align with the
web editor.

It is worth noting this implementation does not match the web
implementation exactly. The web leverages CSS sibling selectors to hide
the writing prompt when an empty paragraph's previous sibling is also
empty. The native mobile implementation cannot easily achieve the same
strategy. So, we instead hide the prompt if the block is empty and not
selected to achieve similar, but slightly different results.