Reorder resets / cleanup.
Resets are things like deleting posts, removing a filter/action, etc.

These need to occur _before_ an assertion.

Why?

If a reset is after an assertion and any of the assertions before it
fail, the test method bails out and the reset does not happen.

To ensure resets happen, they need to be located in either:
* tear_down() fixture.
* just after invoking the thing under test but before the assertion(s).