Don't move focus within the toolbar if it is already focused
If the toolbar has focus already, we don't want to run a requestanimationFrame to then move focus to the previously unmounted index. Adding a check to see if focus is within the toolbar already, then not running the RAF fixes an issue where moving the block via the mover arrows would steal focus to the index before mount.

Also added a fix for getAllFocusableToolbarItemsIn to not exclude aria-disabled buttons, as those can still receive focus.