Reverts first attempt and now tries to harmonize the editor canvas is-focus-mode between site and post editors. By "harmonize", I mean use the same approach by driving the padding via a classname `is-focus-mode` on the edit-${site/post}-layout element, not centralize the code. The latter is difficult because the site editor has "modes", such as edit and view, which also affect the styles.
Reverts first attempt and now tries to harmonize the editor canvas is-focus-mode between site and post editors. By "harmonize", I mean use the same approach by driving the padding via a classname `is-focus-mode` on the edit-${site/post}-layout element, not centralize the code. The latter is difficult because the site editor has "modes", such as edit and view, which also affect the styles.