Ensure Query Loop Block `queryId` Uniqueness
There are multiple cases which can cause `useInstanceId()` to return
IDs that are already in use. Since this function returns IDs based on
the number of given blocks present, we will just set the `queryId`
to the `instanceId` on every component mounting. This approach
is a little bit unstable but it's good enough for the usage.