Fix broken list markup in navigation block when 3rd party blocks are used as decendants of navigation block (#55551)
This introduces a new filter called `block_core_navigation_listable_blocks` that allows extenders to modify the list of blocks that should get wrapped within a `li` element when nested inside the Navigation block. 

On top of that it also modifies the mechanism that controls the `ul` markup to no longer check for a hardcoded list of blocks that are list items. But instead uses the `WP_HTML_Tag_Processor` to check the actual markup to see whether something is a list item.

Co-authored-by: fabiankaegy <fabiankaegy@git.wordpress.org>
Co-authored-by: getdave <get_dave@git.wordpress.org>
Co-authored-by: scruffian <scruffian@git.wordpress.org>
Co-authored-by: draganescu <andraganescu@git.wordpress.org>
Co-authored-by: ndiego <ndiego@git.wordpress.org>
Co-authored-by: justintadlock <greenshady@git.wordpress.org>