Privately export useHasAnyBlockControls for use with top toolbar
There are times where a block is selected but no tools are available. If we don't have any tools, we don't want to show a toolbar. This privately exports the same check used by the block toolbar so we can have a more reliable check if we should show the block toolbar or not.