Refactor for clarity
I debated between setting the isBlockToolsCollapsed within the useEffect or not. I decided to set it manually in the edit-mode__center because the isBlockToolsCollapsed is set by the user, and having an a user defined toggle alongside all of our own settings gets really complicated really quickly. I think it's best to keep that as its own state, and then have alternative checks before we honor that setting.

This also adds a fix where the .is-collapsed class was added to the central edit-mode__center when:
- top toolbar on
- click a block
- turn off top toolbar
- is-collapsed class would be present (but not honored) because of wrapping CSS