test: Avoid duplicate call to mock internal `TextInputState` module
We already invoke `jest.mock()` for this module within the Jest
environment setup file. We do so as _not_ mocking this module causes
widespread failures in tests due to its relationship to native modules.
So, in a way, it makes sense to globally mock it.

For whatever reason, attempting to invoke `jest.mock()` twice for this
module stopped working after upgrading to React Native 0.73. The mock
function passed to the second invocation was ignored.

This change now updates the existing mock instead.