Remove repository specific Code of Conduct.
This removes the repository specific `CODE_OF_CONDUCT.md` file in favor of the newly added version maintained as the organization's default within the WordPress/.github repository.

While this file will not be included in `git clone`s, this change enables more consistent messaging for the community code of conduct, and allows the file to be updated in one location for all repositories under the WordPress organization.