refactor: Align `RichText` refs with web implementation
Previously, the native `RichText` implementation relied upon a bespoke
`setRef` prop to forward refs to the native rich text editor instance
(i.e., Aztec). While this worked fine for contexts where `RichText` was
used for the native platform only, it diverged from the web `RichText`.
This divergence created incompatibility for the context of a component
running on both web and native which passed a ref to `RichText`.

This refactor aligns the native `RichText` with the web implementation
so that refs may be passed under a single, unified `ref` prop. The ref
is then forwarded onto the native rich text editor instance, as it was
before.