Scale zoom out mode frame based on available content width
When there's a small area available to scale the editor, we want to have the editor larger. This implements a smooth sloped clamped scale to scale the zoomed out editor to better fit its available space.

We moved the scale calculation inside of the iframe components so when this is used within the edit post and other packages, they won't need to implement their own scaling.