Use getBlockSelectionStart to check for selectedBlockClientId
selectedBlockClientIds() does quite a bit more work, when we only need one of the selected block ids. Generally, this will only be one block id. In the case of multiples, the previous method would get the topmost block in the selection in the array. getBlockSelectionStart gets the block where selection starts, which is a slightly different concept. Functionally, I think the outcome is identical, and getBlockSelectionStart should be faster.