Data Views: Ignore cmd-click when row not selectable
In `ViewGrid`, move the check for `hasBulkAction` out of the condition
and into the block of the outer `if` statement. This is done so that the
event mechanics remain the same whether the row is selectable or not;
in particular, the statements to event stop propagation and prevent
default shouldn't be predicated on that ability.