Data Views: Ignore cmd-click when row not selectable (#59697)
When using `cmd + click` on an item that isn't selectable, simply ignore
that event, thereby preventing the item's checkbox from flashing from
checked to unchecked. This was already the case in the Grid view, and
this applies the same idea to the Table view.

In `ViewGrid`, move the check for `hasBulkAction` out of the condition
and into the block of the outer `if` statement. This is done so that the
event mechanics remain the same whether the row is selectable or not;
in particular, the statements to event stop propagation and prevent
default shouldn't be predicated on that ability.