Add a clickDelay via a 100ms setTimeout to prevent the link control ui from opening again.
We can't toggle the addingLink state via the link because the onFocusOutside event fires to close the popover, and a the click on the inline link opens the link control again. A better fix to this issue would be to have the onFocusOutside event report the correct element that was clicked, and we could check the element that was clicked.