Re-introduce `data_wp_context()` with `_deprecated_function()` call
Following its removal in https://github.com/WordPress/gutenberg/pull/59465 and release in Gutenberg 17.9, this can potentially result in fatal errors on sites where plugins or themes are calling `data_wp_context()` instead of the newer function name.

This should backfill it in, and provide a reminder to update to the newer function.