fix: Avoid crash by preventing quick src changes
There is a bug with `react-native-video` that causes a crash if a video's `src` is changed quickly after setting it with some other value. Relevant GitHub discussion:  https://github.com/wordpress-mobile/WordPress-iOS/issues/20882#issuecomment-1598228014

With this commit, we prevent the `src` from being quickly changed by not showing a video if it still has the `file:` protocol.