Interactivity: Ensure stores are initialized on client (#59842)
If a store call happens after interactivity has hydrated, subscriptions may not be made correctly which means that when a store is later added, the subscriptions do not exist and the client will not update. This is undesirable.

We fix this by creating empty stores during directive resolution, ensuring subscriptions can be made so that subsequent store changes will update the client.