feat: Enable setting HTML anchor IDs for all supported block types (#59802)
* feat: Support inspector advanced controls

Allow blocks displaying "advanced" controls within a separate inspector
section.

* feat: Enable block anchor support

Expand anchor support to include all blocks, not just the Heading block.

* docs: Add change log entry

* test: Fix hidden block setting button

Swap the Image block for the Shortcode block for the subject of a block
with no settings now that anchor support is present for an empty Image
block.

* test: Assert setting anchor HTML ID

* refactor: Remove web-specific properties

There is no reason to retain this unused code in the native-specific
file.

Unlinked contributors: iamthomasbishop, jbinda.

Co-authored-by: dcalhoun <dpcalhoun@git.wordpress.org>
Co-authored-by: jhnstn <jsnjohnston@git.wordpress.org>
Co-authored-by: derekblank <derekblank@git.wordpress.org>
Co-authored-by: pinarol <wppinar@git.wordpress.org>