Refactor Enter keypress on Nav Link (#59848)
* Remove unnecessary Enter keypress check onKeydown

Enter is already a key handled by onClick, so this Enter onKeydown never runs

* Remove impossible to reach onClick state

The rich text element had an onClick that would only run if ! url. The wrapping render check only outputs that rich text element if a url is availabe, so this onClick check can never be true.