Media & Text: Make the alt text editable for featured images
Remove the featuredImageAlt constant and only use the mediaAlt constant for the alt text.
This caveat about this is that the alt text that is set for the featured image in the media library is never used in the Media & text block.

When the setting "Use featured image" is toggled, reset mediaAlt attribute to an empty string.
If the alt text is changed through the option in the block settings panel, update the mediaAlt attribute.

The changes to index.php ensures that:
The img tag always has an alt attribute, even if the alt attribute is empty.
When there is a custom alt text saved to the attribute, it is used.