Fix zoom out mode toggling between pattern category selection (#60225)
* Move useZoomout check up to PatternCategeryPreviewsPanel wrapper

Previously, the useZoomOut check was within the category previews rather than the panel. This meant when the category changed, the zoom out check ran again. We only want the zoom out mode check to happen when the panel itself is mounted.

* Simplify useZoomOut logic

The useRef wasn't necessary, as it was running before the other useEffect that could set it to true.