Remove shouldUseKeyboardShortcut from DocumentTools
shouldUseKeyboardShortcut was toggling to say if the NavigableToolbar within DocumentTools should use the alt+f10 shortcut based on if a toolbar was available or not. This was causing the Header to rerender everytime, which was performance drag. Removing this means the Header won't rerender as often, and the test coverage for the alt+f10 shortcut still all works as intended, so this prop is not necessary to toggle.