Remove block-editor package usage from components (#60999)
Remove references to @wordpress/block-editor from @wordpress/components package.

This fixes an issue where dependencies crossed TypeScript project boundaries triggering errors like:

> error TS5055: Cannot write file '…/alignment-matrix-control/icon.d.ts' because it would overwrite input file.

This is because the block-editor project and the components project both attempt to write the same files because project references do not reflect this interdependence. In fact, block-editor depends on components. That would create a circular dependency between the projects which TypeScript does not allow.

---------

Co-authored-by: sirreal <jonsurrell@git.wordpress.org>
Co-authored-by: mirka <0mirka00@git.wordpress.org>
Co-authored-by: jsnajdr <jsnajdr@git.wordpress.org>
Co-authored-by: geriux <geriux@git.wordpress.org>