build: Suggest workaround if `tsc --build` fails
Following the merge of #60796, developers may face build issues that
require package types to be rebuilt. The problem is that `tsc --build`
fails somewhat silently -- or rather, there is some output, but it's not
clear in the console which stage of the `build:package-types` command
failed, and hence what the workaround should be.

This commit alleviates the issue by logging a helpful message in the
console if `tsc --build` fails:

    tsc failed. Try cleaning up first: `npm run clean:package-types`