Clear last focused block when inserting new blocks
When inserting a new block from the inserter, focus should go to this block when focus is moved to the canvas via the Tab keypress. To do this, we need to clear where focus previously was when a new block is inserted, otherwise it tries to place focus where it was before.