build:package-types: Run silently to reduce user confusion
Follows up on #61501.

The original PR, #61501, modified the `build:package-types` script to
log a help message to developers if the `tsc --build` call failed. This
message suggested a fix (to run `npm run clean:package-types`).

This works, but due to the way that NPM logs the execution of successive
NPM scripts, the code that conditionally outputs the help message is,
itself, always output. This is not only noisy, but could mislead
developers into following the help message's instruction unnecessarily.

This commit fixes the problem by wrapping the original NPM script in a
new script, which is then called with NPM's `--silent` flag. The console
output now looks like this:

> gutenberg@18.3.0 build:packages
> npm run build:package-types && node ./bin/packages/build.js

> gutenberg@18.3.0 build:package-types
> npm run --silent build:package-types:verbose