fix: improve dataview types
This commit improves the types in the dataviews package.

This is helpful by typing the `item` or `items` provided as arguments
to the functions of the fields and actions of the `DataViews`.

This commit does not implement requiring these types in the `DataView`
component, but helps to improve the types of the functions that are
used in the `DataView` package. It also allows exporting the types
to be used in consumers of the package.

- Add generic item type to all dataview types that operate
on `item` or lists of `items`.

- Update the usage in the dataviews package to incorporate
the use of the generic item type.

- Add an `Action` type for the dataview actions.