Add const assertion to sortValues
When TypeScript infers the type of an object literal like this:

{ key: 'value' }

It infers that the value is of type `string`. A const assertion tells
TypeScript it can treat the literal as a constant, so instead of
`string`, the type of the value can become the string literal `'value'`.

This fixes an issue where `string` would not satisfy a type that
expected a union of literal values. With the const assertion, TypeScript
knows that e.g. the type of the value `sortValues.asc` is `'ascending'`
instead of `string`.