Enable shadow support for cover block
On previous attempts to do this, the cover block was displaying
inconsistencies between how it was rendered in the editor vs the
frontend; that doesn't seem to be the case anymore, so we should be
able to enable shadow support without any modifications to the block itself.

I'd still appreciate some eyes on this, because I don't know what
exactly has changed since the first time I attempted this.

Tracking issue: https://github.com/WordPress/gutenberg/issues/57103

Co-authored-by: vcanales <vcanales@git.wordpress.org>
Co-authored-by: madhusudhand <madhudollu@git.wordpress.org>